feat(game): add UUID to game configuration
This commit introduces the addition of a UUID to the game configuration, allowing for unique identification of game instances. Moreover, modifications have been made in the initialization of the game to utilize the provided UUID instead of generating it internally. Additionally, the changes for setting the maximum consumers per client in the message publisher have also been included in this commit.